requirements.txt
faster-whisper>=1.0.0
torch>=2.0.0
numpy>=1.24.0
loguru>=0.7.0
pyyaml>=6.0
httpx>=0.24.0
tenacity>=8.2.0
prometheus-client>=0.17.0
soundfile>=0.12.0
librosa>=0.10.0
groq>=0.4.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
websockets>=11.0
jsonschema>=4.17.0
